cpp: Move programs to libexec to avoid path clashes
