+++ CMakeLists.txt
@@ -77,8 +77,12 @@ set(OME_BIOFORMATS_INSTALL_FULL_DATADIR "${CMAKE_INSTALL_FULL_DATADIR}/${OME_BIO
 
 # Use standard path if using a prefix.
 if(CMAKE_INSTALL_PREFIX)
+  set(OME_BIOFORMATS_INSTALL_LIBEXECDIR "${CMAKE_INSTALL_LIBEXECDIR}/ome/bioformats")
+  set(OME_BIOFORMATS_INSTALL_FULL_LIBEXECDIR "${CMAKE_INSTALL_FULL_LIBEXECDIR}/ome/bioformats")
   set(OME_BIOFORMATS_ICONDIR icons/hicolor/scalable)
 else()
+  set(OME_BIOFORMATS_INSTALL_LIBEXECDIR "${CMAKE_INSTALL_LIBEXECDIR}")
+  set(OME_BIOFORMATS_INSTALL_FULL_LIBEXECDIR "${CMAKE_INSTALL_FULL_LIBEXECDIR}")
   set(OME_BIOFORMATS_ICONDIR icons)
 endif()
 set(OME_BIOFORMATS_INSTALL_ICONDIR "${CMAKE_INSTALL_DATADIR}/${OME_BIOFORMATS_ICONDIR}")
