+++ cpp/libexec/showinf/ImageInfo.cpp
