+++ cpp/libexec/showinf/ImageInfo.h
