+++ cpp/libexec/showinf/options.cpp
