+++ cpp/libexec/showinf/options.h
