+++ cpp/libexec/showinf/showinf.cpp
